+++ source/extensions/omni.usd.schema.isaac/python/__init__.py
@@ -15,5 +15,6 @@ pluginsRoot = os.path.join(os.path.dirname(__file__), "../../../plugins")
 
 Plug.Registry().RegisterPlugins(pluginsRoot + "/DrSchema/resources")
 Plug.Registry().RegisterPlugins(pluginsRoot + "/RangeSensorSchema/resources")
+Plug.Registry().RegisterPlugins(pluginsRoot + "/IsaacSensorSchema/resources")
 Plug.Registry().RegisterPlugins(pluginsRoot + "/RobotEngineBridgeSchema/resources")
 Plug.Registry().RegisterPlugins(pluginsRoot + "/RosBridgeSchema/resources")