+++ source/extensions/omni.usd.schema.isaac/docs/CHANGELOG.md
@@ -1,8 +1,16 @@
 **********
 CHANGELOG
 **********
+[0.2.0] - 2022-03-21
+========================
 
+Added
+-------
+- Add Isaac Sensor USD Schema 
+- Add Contact Sensor USD Schema
+- Add IMU Sensor USD Schema
 [0.1.1] - 2021-08-02
+
 ========================
 
 Added